ok, fixed spacing issues so that we don't replace every space with a &nbsp; because if we did, it produced non-wrapping lines.
I've got it checking to see if there are 3 spaces in a row, and replacing them
with "&nbsp; &nbsp;" and else if there are 2 spaces in a row, "&nbsp; "  I
think this should address all problems there may be.  Let me know if you can
think of a situation this wouldn't account for.  ok, thanks. bye :)

-pav