Delete works.
*NOTE* BE CAREFUL:
It does *not* undelete (yet).  So if you mark a message read, you had better
not exit through a normal means.. (so kill -9 it or something:)  or just be
careful.

It will (should) purge all messages marked delete upon exit or change in
mailbox.  I'll put a better way to do this after i've done undelete.

-pav