bleh, :)
take a look at libbalsa/mbox.c if you would like to help.. it needs to read
through the file, and put the headers for each message into the message
structure.. when that message is done, add it to the glist.  don't forget to
set mailbox->offset with the offset of each message.  at the end, return the
glist...  no, it won't work if you remove the #if 0's, but if someone puts
something together, i'll make it work.
-pav