I completely rewrote the configuration management stuff. It uses
I completely rewrote the configuration management stuff.  It uses
libPropList now and saving and loading work.  It needs tweaking to
set which configuration options are optional and which are mandatory
and etc etc, but it's mainly there.

You might want to update your libPropList if you're going to use this,
since I found a bug in it which I fixed.

I'd also like to say that it would be nice if Balsa built -Wall, so someone
should see about turning that on.

The (woefully incomplete) ChangeLog entry follows.

1998-08-23  Nat Friedman  <ndf@retrans.mit.edu>

	* src/save-restore.c: This file has been completely rewritten.
	libPropList based configuration loading and storing is now
	functional.

	The configuration function names have all been made sane, also,
	and several files, including balsa-app.c, balsa-init.c, main.c,
	and pref-manager.c had to be modified to call the new routines.

	* src/mailbox-conf.c: Added some casts to the gtk_signal_connect()
	calls so that passing non-pointer closures doesn't generate a warning.