Reworked the way in which LibBalsa checks mailboxes for new mail.
This involved adding a virtual function to LibBalsaMailbox
which is implemented by each mailbox to check itself, and
a signal to be emitted when a mailbox toggles it's
have_new_messages flag.

Also added notify.[ch] which encapsulate the functionality
of buffy and is used to check closed mailboxes.

Updated the balsa application code to take advantage of
this new checking code.