Silence a gcc gripe about strncpy
when the byte count equals the buffer size. Copy at most one fewer bytes,
so that gcc *knows* that we are going to nul-terminate the destination
string. This makes no operational difference, as the last byte would
be overwritten with the nul in the exceptional case when the limit
is reached.

	* libbalsa/imap/imap-commands.c (imap_mbox_handle_fetch_body):
	silence a gcc gripe about strncpy.