mailbox: Sort by thread date
The 'thread-date' of a message is the date of the most recent message in
the thread that it heads. This commit changes 'sort by date' to 'sort by
thread-date'.

It does not change the displayed date of a message, which continues to
be its real date. But an old thread with a recent reply is sorted by the
date of the reply, thus avoiding the hiding of recent messages in old
threads.