address-book-rubrica: Do not crash on bad card
Do not crash if a card has multiple "Net" sections.

* libbalsa/address-book-rubrica.c (extract_cards): Ignore all
  but the first non-empty "Net" section.

If we call libbalsa_address_set_addr_list() and then call it again with
either the same list or an extended list, we free all the e-mail
addresses from the first call, and then free the list before setting it.
Only bad things can follow!