mime-widget-text: Connect to the "pressed" signal
of a GtkGestureMultiPress, instead of the "button-press" widget event.
And set the gesture's button to 0.

The gesture and the popover should be attached to the web-view, not our
LibBalsaHtml container. If the info-bar is not showing there is no
difference, but when it is showing it should not pop up the context
menu; also, the event coordinates are relative to the web-view, not the
container.