libbalsa: In popup-widget, special case GtkToolbar
When the user right-clicks on a toolbar button, the GdkEvent coordinates
are relative to the button. The root x-coordinate is relative to the
toolbar, so we use that to position the popup, instead of the event x.

The event y-coordinate is a few pixels off from y relative to the
toolbar, and making the pointed-to rectangle 5x5 instead of 0x0 mostly
fixes it. The root y-coordinate is way off, and it seems better to use
the event y.