balsa-index: pop up the context menu on the left
The default position is above the pointer, but that is often infeasible,
so it is actually popped below. But when a submenu like "toggle" or
"move" is selected, it pops up above, if feasible, which is not a
smooth transition. If we choose on the left, there seems to be much less
chance that it is infeasible, so the submenus transition smoothly. That
may depend on the panes layout, for which I'm using the default. We
could perhaps use a layout-dependent choice.