Various: Drop libbalsa_{radio,toggle}_activated
and replace them by NULL in GActionEntry structs.

https://developer.gnome.org/gio/stable/GActionMap.html#GActionEntry:
Since GLib 2.40, [activate] can be NULL for stateful actions, in which
case the default handler is used. For boolean-stated actions with
no parameter, this is a toggle. For other state types (and parameter
type equal to the state type) this will be a function that just calls
change_state…

which is exactly what our helpers did.

* libbalsa/application-helpers.c (libbalsa_window_add_accelerator):
* libbalsa/application-helpers.h:
* libbalsa/source-viewer.c (libbalsa_show_message_source):
* src/ab-main.c (get_main_menu):
* src/main-window.c:
* src/message-window.c:
* src/sendmsg-window.c: