workaround for broken message index heights
Calculate the balsa message index cell renderer height as fixed value
from the maximum of 1.25 times the standard font height and the menu
size icon height as to work around the automatic calculation being
confused by rows using bold and/or italics.  Note: not used if
TREE_VIEW_FIXED_HEIGHT is not defined.

Signed-off-by: Albrecht Dreß <albrecht.dress@arcor.de>