mime-widget-text: Draw cite bars in an idle cb
We used to use the "draw" signal, so that the text-view had time to
compute the locations of cited text, but that has gone away. Neither the
"realize" nor "map" signals comes late enough, so we wait for "realize"
then schedule an idle callback, and that seems to be long enough.