Check for GLIB stable version
g_memdup() will be deprecated in version 2.68, so check for that version
before using g_memdup2().

The only places where it is called are with a fixed small byte_size
argument that cannot overflow when cast from size_t to unsigned, so
the code is not vulnerable; we need only to avoid the deprecation
warning.

modified:   libbalsa/mailbox_mbox.c