Autocrypt: fix faulty error messages
Fix faulty Autocrypt error messages (false positives) which occur on the
following conditions:

(1) As g_mime_message_get_autocrypt_header() will return an object even
if the message does *not* include an Autocrypt: header,
balsa_mime_widget_signature_widget() will report an error for a signed
message without such a header, but no key being available.  This is
fixed by remembering if the message actually contains the header or not.
(2) The Autocrypt key material must include a signing-capable subkey,
but it is not necessarily the first subkey.  In the latter case, the
wrong fingerprint is stored in the database.  Note that the user must
manually remove the entry from the Autocrypt database to get rid of the
faulty error in this case.

Details:
- libbalsa/message.[ch]: remember if the message contains an Autocrypt:
header
- libbalsa/autocrypt.c: loop over all Autocrypt subkeys, and return the
fingerprint of the first signing subkey instead of the first subkey
- src/balsa_mime_widget_crypto: emit error message about a broken
Autocrypt header iff the message actually contains it.

Signed-off-by: Albrecht Dreß <albrecht.dress@netcologne.de>