mailbox_mh: Use g_byte_array_new()
instead of casting g_array_new() to GByteArray*. We did this to make the
data zero-terminated, but it's undocumented (and ugly). We already
appended a zero byte elsewhere in mailbox_mh, so just do the same here.
Also, define "static const guint8 zero[1] = {0}", to match the signature
of g_byte_array_append().