mime: No trailing comma in a URL
The regex that Balsa uses to identify a URL in a plain text message part
already omits a trailing period ('.') from the match, as it is usually
punctuation and not part of the URL. This commit additionally removes a comma
(',') from the last character match, for the same reason: more likely to be
punctuation. This of course breaks the identification of a URL that *does* end
in a comma, but that appears to be less likely than erroneously including one.

Discussion at https://gitlab.gnome.org/GNOME/balsa/-/issues/59.