balsa-mime-widget-text: Hide quote prefix
Revert the previous commit and try a different approach. As Albrecht pointed out at
https://gitlab.gnome.org/GNOME/balsa/-/merge_requests/45#note_1286766, it resulted
in copied text not containing the quoting prefix, which is an undesirable change
in behavior.

We can fix the problem by using gtk_text_view_get_line_yrange() instead of
gtk_text_view_get_iter_location(). The former looks across the whole line, whereas
the latter looks at only a single character; if it is one that has been made smaller
to hide it, the location and height are not what we want.