balsa-mime-widget-text: Hide quote prefix
A recent GTK change led to `Balsa`'s cite-bars being misrendered. We can fix the problem by using `gtk_text_view_get_line_yrange()` instead of `gtk_text_view_get_iter_location()` to find the correct locations of the top and bottom of a bar. The former looks across the whole line, whereas the latter looks at only a single character; if it is one that has been made smaller to hide it, the location and height are not what we want.