send: Drop unused parameter
The "reply_message" parameter of libbalsa_message_postpone() is unused.
Remove it, and fix the calls.

When replying to a message, Balsa flags it as "replied" when the reply
is sent. If the reply is postponed, the reference to the original
message is dropped, so it cannot be automatically flagged. Hence passing
a pointer to the original message is pointless.

modified:   libbalsa/send.c
modified:   libbalsa/send.h
modified:   src/sendmsg-window.c