Move declaration of libbalsa_wrap_quoted_line()
from libbalsa/mime.h to libbalsa/misc.h.

It belongs in libbalsa/misc.c, but defining it there puts it in misc.o,
and linking balsa-ab drags in misc.o, which then pulls in mime.o to
satisfy the reference to libbalsa_match_regex, and it goes downhill
from there.

So we define it in mime.c, to keep misc.o clean, but declare it in
misc.h, with the other wrapping functions.