HTML message display and printing improvements
See Gitlab issue #75 for further details.

Basically, the following is implemented in this branch:
(1) remember for each multipart/alternative if the plain or html part
has been selected by the user for display,
(2) remember for each html part if the user decided to load external
contents, or if it has been loaded due to a rule,
(3) if a message if opened in a 2nd window display it exactly as in the
1st view,
(4) print the message exactly as displayed in the view from which
printing is initiated, and remove all related config items from the
print dialogue.

Details:
- libbalsa/body.[ch]: add flags and a reference to the parent part to
_LibBalsaMessageBody plus helper setter/getters for (1) and (2);
- libbalsa/html.[ch]: use the _LibBalsaMessageBody flag re. already
loaded external content;
- libbalsa/mime.c, src/balsa-mime-widget-text.c: clarify the use of the
LibBalsaHTMLType enum
- src/balsa-message.c: set/use html selection status, clarify the use of
the LibBalsaHTMLType enum
- src/balsa-print-object-html.c: modified html print api
- src/balsa-print-object.h: remove needless html print config items
- src/print-gtk.c: rework selection of parts to be printed, remove html
print config items

Signed-off-by: Albrecht Dreß <albrecht.dress@netcologne.de>