meson.build: Be careful with run_command()
See https://github.com/mesonbuild/meson/issues/9300

In some future version of meson, the `check` kwarg of `run_command()`
will default to `true` instead of `false`, meaning that `meson` will
error out if the command fails. When we check `.returncode()`, we don't
want that to happen, so explicitly set `check: false`. If we don't use
`.returncode()`, we expect the command to succeed, so allowing `meson`
to error out should be desirable.