balsa-mime-widget-callbacks: Handle empty parts
When saving a message part, write it to a temporary file. If an error
occurred or if zero bytes were written, the temporary file doesn't exist.
Otherwise, move it to the chosen destination. This preserves an existing
file when there's nothing to replace it.

See https://gitlab.gnome.org/GNOME/balsa/-/issues/64.