sendmsg-window: Clean up BalsaSendmsg
Replace all GtkWidget header pairs with:

    struct _BalsaSendmsgHeader {
        GtkWidget *name;
        GtkWidget *body;
    };

It makes no functional change and no change in the size of BalsaSendmsg,
but it makes the code more readable and maintainable.

BalsaSendmsgHeader:name is always a GtkLabel; BalsaSendmsgHeader:body may be
a GtkEntry, a GtkComboBox, or a GtkFrame containing a LibBalsaAddressView,
depending on the header.

modified:   src/ab-window.c
modified:   src/sendmsg-window.c
modified:   src/sendmsg-window.h