html: fix processing missing or empty “cid:<content-id>” data
Fix 1: libbalsa_message_body_get_content() may return a non-NULL buffer
even if its length is 0, which in this case is not freed.
Fix 2: if the referenced content part doesn't exists in the (broken)
message, the URISchemeRequest is not terminated.  Whilst this is (for a
reason I don't understand) apparently harmless for a display view, it
blocks html printing, and the hanging WebKitWebProcess child prevents
Balsa from terminating cleanly.