fix composing a message for a file URI from the command line
The function sw_attach_file() in src/sendmsg-window.c expects a file
name for creating a message for a file from the command line.  This
works for `xdg-email`, but _not_ for e.g. `thunar-sendto-email` which
passes an URI.  The change checks for a `file://` URI and converts it
into a file name.

Signed-off-by: Albrecht Dreß <albrecht.dress@posteo.de>