remove GpgME passphrase callback
Since the introduction of gpg2 almost two decades ago, the gpg-agent is
generally used for reading passphrases.  Thus, Balsa's GpgME passphrase
callback will never be called in any real-life installation and can be
removed, resulting in a simplification of the code.

During startup, a test accessing the agent through the assuan protocol
is added.  If it fails, the user is warned that crypto operations may
not succeed.

Closes #103 as the xpm icon for the internal passphrase callback is
removed.

Removed API calls:
- lb_gpgme_passphrase() and all dependent stuff

Simplified API calls:
- libbalsa_gpgme_init(): remove callback
- libbalsa_gpgme_new_with_proto(): remove callback and parent window
- libbalsa_gpgme_decrypt(), g_mime_application_pkcs7_decrypt_verify(),
  g_mime_gpgme_mpe_decrypt(), g_mime_part_rfc2440_decrypt(),
  libbalsa_body_decrypt(), libbalsa_rfc2440_decrypt(): remove parent
  window

Signed-off-by: Albrecht Dreß <albrecht.dress@posteo.de>